Move sudo tailing script tailfilreader.rb to omsagent owned /opt/microsoft/omsagent/plugin/ folder
