+++ installer/conf/sudoers
@@ -26,7 +26,7 @@ omsagent ALL=(ALL) NOPASSWD: /opt/microsoft/omsagent/bin/service_control
 omsagent ALL=(ALL) NOPASSWD: /usr/bin/docker inspect *
 
 #run tailfilereader.rb
-omsagent ALL=(ALL) NOPASSWD: /opt/microsoft/omsagent/ruby/bin/ruby /opt/microsoft/omsagent/bin/tailfilereader.rb *
+omsagent ALL=(ALL) NOPASSWD: /opt/microsoft/omsagent/ruby/bin/ruby /opt/microsoft/omsagent/plugin/tailfilereader.rb *
 
 #Enable check for if file exists and is readable
 omsagent ALL=(ALL) NOPASSWD: /usr/bin/test -r *
