+++ source/code/plugins/tailfilereader.rb
